Fix an iOS warning

+++ _sokol_gfx_metal.impl.h
@@ -1182,7 +1182,7 @@ _SOKOL_PRIVATE void _sg_create_pipeline(_sg_pipeline* pip, _sg_shader* shd, cons
                 id<MTLFunction> mtl_vs_func = _sg_mtl_pool[shd->stage[SG_SHADERSTAGE_VS].mtl_func];
                 for (MTLVertexAttribute* mtl_attr in mtl_vs_func.vertexAttributes) {
                     if (0 == strcmp(mtl_attr.name.UTF8String, attr_desc->name)) {
-                        mtl_attr_index = mtl_attr.attributeIndex;
+                        mtl_attr_index = (int)mtl_attr.attributeIndex;
                         break;
                     }
                 }
